reference-code/backend-v2/pyproject.toml
[project]
name = "backend-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "e2b>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]
